test(rsc): split independent tests
